Add files via upload
seasonally adjusts CAGED data using a bottom-up approach.  

this one is a little complicated. since the CAGED is made availabe by MTE via an excel spreadsheet, I have a VBA script (also in this repository) that treats it for me and creates a .csv file ready to be read by this R script. the R script then separates the data by sector and by what i've called "subsector", seasonally adjusts both using the "seas" package and creates an excel file for each.